Move all non-library code outside of the imapclient package
The distribution will only include the code that it actually
needed to use the library. Unit tests, live tests and examples
are moved to the root of the repository, like on most other
Python projects.

Fixes #266
