+++ examples/idle_example.py
