+++ tests/test_starttls.py
@@ -4,7 +4,7 @@
 
 from __future__ import unicode_literals
 
-from ..imapclient import IMAPClient
+from imapclient.imapclient import IMAPClient
 from .imapclient_test import IMAPClientTest
 from .util import Mock, patch, sentinel
